requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
pillow
numpy
